pyproject.toml
[project]
name = "toolbx-containers"
version = "0.1.0"
description = "Containers meant to be used by Toolbx."
readme = "README.md"
requires-python = ">=3.13"
dependencies = ["invoke>=2.2.0", "python-dotenv>=1.1.1"]

[tool.ruff]
line-length = 99
src = ["tasks", "build"]
# Ruff will automatically exclude all files listed in .gitignore as well as common temporary Python
# tool directories.
# To exclude additional folders, use extend-exclude.

[tool.ruff.lint]
select = [
    "F",   # pyflakes
    "E",   # pycodestyle
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "RUF", # ruff
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "ISC", # flake8-implicit-str-concat
    "PIE", # flake8-pie
    "PT",  # flake-pytest-style
    "PTH", # flake8-use-pathlib
    "SIM", # flake8-simplify
    "TID", # flake8-tidy-imports
]
extend-ignore = ["RUF005", "RUF012"]
unfixable = [
    # Disable removing unused imports by default and only enable within nox so editors don't delete
    # unused imports while the user is in the middle of editing a file on save.
    "F401",
]

[tool.ruff.lint.isort]
force-sort-within-sections = true
split-on-trailing-comma = false
# For non-src directory projects, explicitly set top level package names:
# known-first-party = ["my-app"]

[tool.ruff.lint.flake8-tidy-imports]
ban-relative-imports = "all"
